+++ app/src/main/java/dk/scheduling/schedulingfrontend/ApiButton.kt
@@ -1,14 +1,9 @@
 package dk.scheduling.schedulingfrontend
 
 import android.util.Log
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Button
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.sp
 import dk.scheduling.schedulingfrontend.api.ApiClient
 import retrofit2.Call
 import retrofit2.Callback
