+++ app/src/main/java/dk/scheduling/schedulingfrontend/HomePage.kt
@@ -1,0 +1,16 @@
+package dk.scheduling.schedulingfrontend
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun HomePage() {
+    Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
+        Text(text = "This is the HomePage, click the heart to reach the api button!", fontSize = 24.sp)
+    }
+}
