+++ app/src/main/java/dk/scheduling/schedulingfrontend/MainActivity.kt
@@ -2,17 +2,10 @@ package dk.scheduling.schedulingfrontend
 
 import App
 import android.os.Bundle
-import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Text
-import dk.scheduling.schedulingfrontend.api.ApiClient
 import dk.scheduling.schedulingfrontend.ui.theme.SchedulingFrontendTheme
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 
 class MainActivity : ComponentActivity() {
     @OptIn(ExperimentalMaterial3Api::class)
@@ -20,7 +13,7 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             SchedulingFrontendTheme {
-                //ApiButton()
+                // ApiButton()
                 App()
             }
         }
