+++ app/src/main/java/dk/scheduling/schedulingfrontend/Page3.kt
@@ -1,0 +1,16 @@
+package dk.scheduling.schedulingfrontend
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun Page3() {
+    Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
+        Text(text = "Page 3", fontSize = 24.sp)
+    }
+}
